move sharing contracts
