+++ src/backend/Application/Features/Sharing/Endpoints/RevokeSharedLinkEndpoint.cs
@@ -1,44 +1,39 @@
 using Application.Extensions;
 using FastEndpoints;
 
-namespace Application.Features.Sharing.Endpoints;
-
-public record RevokeSharedLinkRequest
+namespace Application.Features.Sharing.Endpoints
 {
-    /// <summary>The shared link id (bound from the route).</summary>
-    public string LinkId { get; set; } = null!;
-}
-
-/// <summary>
-/// Revokes a shared link. Requires authentication. Only the link creator or video owner can revoke.
-/// </summary>
-public class RevokeSharedLinkEndpoint : Endpoint<RevokeSharedLinkRequest>
-{
-    private readonly ISharedLinkService sharedLinkService;
-
-    public RevokeSharedLinkEndpoint(ISharedLinkService sharedLinkService)
+    /// <summary>
+    /// Revokes a shared link. Requires authentication. Only the link creator or video owner can revoke.
+    /// </summary>
+    public class RevokeSharedLinkEndpoint : Endpoint<RevokeSharedLinkRequest>
     {
-        this.sharedLinkService = sharedLinkService;
-    }
+        private readonly ISharedLinkService sharedLinkService;
 
-    public override void Configure()
-    {
-        Delete(ApiRoutes.Share.Revoke);
-        Policies(ApiScopes.Read);
-    }
-
-    public override async Task HandleAsync(RevokeSharedLinkRequest req, CancellationToken ct)
-    {
-        var userId = User.GetSubject();
-
-        var result = await sharedLinkService.RevokeSharedLinkAsync(req.LinkId, userId, ct);
-
-        if (!result)
+        public RevokeSharedLinkEndpoint(ISharedLinkService sharedLinkService)
         {
-            await Send.NotFoundAsync(ct);
-            return;
+            this.sharedLinkService = sharedLinkService;
         }
 
-        await Send.NoContentAsync(ct);
+        public override void Configure()
+        {
+            Delete(ApiRoutes.Share.Revoke);
+            Policies(ApiScopes.Read);
+        }
+
+        public override async Task HandleAsync(RevokeSharedLinkRequest req, CancellationToken ct)
+        {
+            var userId = User.GetSubject();
+
+            var result = await sharedLinkService.RevokeSharedLinkAsync(req.LinkId, userId, ct);
+
+            if (!result)
+            {
+                await Send.NotFoundAsync(ct);
+                return;
+            }
+
+            await Send.NoContentAsync(ct);
+        }
     }
 }
