+++ src/backend/Application/Features/Sharing/Endpoints/VideoInfoBySharedLinkRequest.cs
@@ -1,0 +1,8 @@
+namespace Application.Features.Sharing.Endpoints
+{
+    public record VideoInfoBySharedLinkRequest
+    {
+        /// <summary>The shared link id (bound from the route).</summary>
+        public string LinkId { get; set; } = null!;
+    }
+}
