+++ src/backend/TB.DanceDance.API.Contracts/Features/Sharing/ListMySharedLinksResponse.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using TB.DanceDance.API.Contracts.Features.Sharing;
+
+namespace Application.Features.Sharing.Endpoints
+{
+    public class ListMySharedLinksResponse
+    {
+        public IReadOnlyCollection<SharedLinkResponse> Links { get; set; } = Array.Empty<SharedLinkResponse>();
+    }
+}
