+++ src/backend/TB.DanceDance.API.Contracts/Features/Sharing/RevokeSharedLinkRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace Application.Features.Sharing.Endpoints
+{
+    public class RevokeSharedLinkRequest
+    {
+        /// <summary>The shared link id (bound from the route).</summary>
+        public string LinkId { get; set; } = string.Empty;
+    }
+}
